requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pymongo==3.12.0
bson==0.5.10
pydantic==1.8.2
python-dotenv